Remove hand rolled ohm parser
